beta-ui-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
cachetools==5.5.2
certifi==2024.7.4
charset-normalizer==3.4.1
click==8.2.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.3.1
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.61
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langsmith==0.3.42
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
mysql-connector-python==9.3.0
networkx==3.4.2
numpy==2.1.3
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.0
py4j==0.10.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PySocks==1.7.1
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
safetensors==0.5.3
selenium==4.23.1
sentencepiece==0.2.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
some-package==0.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
zstandard==0.23.0
